+++ src/main/java/br/com/erp/sapx/usecase/projects/UpdateProjectUseCase.java
@@ -4,5 +4,5 @@ import br.com.erp.sapx.domain.Project;
 import br.com.erp.sapx.domain.UpdateProjectRequest;
 
 public interface UpdateProjectUseCase {
-    Project execute(Integer numProjeto, UpdateProjectRequest request);
+    Project execute(Long numProjeto, UpdateProjectRequest request);
 }
