+++ src/main/java/br/com/erp/sapx/usecase/projects/impl/FindProjectByNumProjetoUseCaseImpl.java
@@ -17,7 +17,7 @@ public class FindProjectByNumProjetoUseCaseImpl implements FindProjectByNumProje
     private final ProjectGateway gateway;
 
     @Override
-    public Project execute(Integer numProjeto) {
+    public Project execute(Long numProjeto) {
         return ofNullable(gateway.findProjectByNumProjeto(numProjeto))
                 .orElseThrow(() -> new HttpClientErrorException(HttpStatus.NOT_FOUND, "Projeto n°" + numProjeto + " não encontrado."));
     }
